// src/tests/workEmail/getWorkEmail.generated.ts
/* eslint-disable */
/* @ts-nocheck */
/* prettier-ignore */
/* This file is automatically generated using `npm run graphql:types` */
import type * as Types from '../../generated/types';

import type { JsonObject } from "type-fest";
import gql from 'graphql-tag';
export type WorkEmailQueryVariables = Types.Exact<{
  email: Types.Scalars['String']['input'];
}>;


export type WorkEmailQuery = { __typename?: 'Query', workEmail: { __typename: 'WorkEmail', id: string, isValidated: boolean } };


export const WorkEmail = gql`
    query WorkEmail($email: String!) {
  workEmail(email: $email) {
    __typename
    id
    isValidated
  }
}
    `;